backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
aiohttp
async_timeout
pandas
python-dotenv
tenacity
